Default Changelist

+++ maxlead_site/common/common.py
@@ -166,15 +166,18 @@ def spiders_send_email(f_obj, file_path=None):
     f_obj.close()
     os.remove(file_path)
 
-def kill_pid_for_name(name):
+def kill_pid_for_name(name, type=0):
     lines = os.popen('ps -ef | grep %s' % name)
     for path in lines:
         try:
             client = path.split(' ')
+            if type == 1:
+                client = ['idle']
             if "idle" in client:
                 progress = path.split(' ')[2]
                 if not progress:
                     progress = path.split(' ')[1]
-                os.popen('kill %s' % progress)
+                if progress:
+                    os.popen('kill %s' % progress)
         except:
             continue
